(Binary Analysis) Fixed a flaw in symbolic semantics
When reading a multi-byte value from an uninitialized memory
location, the old version created multiple single-byte variables
and wrote them to memory.  The value returned from the readMemory
was a concatenation of the individual byte variables.  This was
inconsistent with writing a multi-byte value and then reading it,
which returns a single symbolic variable.

The new version, when reading from an uninitialized memory
location, creates a multi-byte variable and then extracts and
writes individual bytes of the variable, which results in the
same expressions being stored in memory as if a multi-byte
write had occurred just before the read.