(Binary Analysis) Bug fix for thunk recognition
When analyzing a function to find a thunk, it is sometimes
possible for the initial JMP to have a target that is outside the
function. This can happen when the function contains other blocks
that are not reachable, but were added to the functionanyway
because they look statistically like code.

(cherry picked from commit 193cf2ab2ae3c482c1096cb68c5cacb13b168eb4 from edg4x repo)