(CMAKE ROSE) abandon convenience library approach
By default, CMake object libraries do not consider themselves "shared"
for the purposes of DEFINE_SYMBOL.  Thus, using these object libraries
on Windows results in undefined external symbols because their symbols
are not exported from the .dll.  For now, we revert to the previous
design of CMAKE ROSE: compile most everything into a single shared
library.  We can revisit this design decision later, if desired.

This commit also contains some support for the
internalFrontendDevelopment configuration option.  It is still untested,
so it probably needs more work.
We can revisit this issue later, if desired.