(ROSE CMAKE) address Jenkins failures
1.  Collapse the generation of preproc-fortran.cc and
preproc-fortran-fixed.cc into a single custom command.  Previously, when
these files were generated by separate custom commands, there existed a race
condition.  This is due to the fact that both files are generated as lex.yy.c
and then renamed appropriately.

2.  Make rosetta_generated depend on generated_rosePublicConfig.
This dependency isn't strictly necessary, but I'm adding it to ensure
that rosePublicConfig.h is generated early in the build process.
A lot of other targets already have a dependency on rosetta_generated,
so I'm just piggybacking off of that to avoid potentially having to modify
a lot of CMakeLists.txt files.