Add reductions support for Nanos
This commit also modifies the tests configuration so, all tests failing
with Nanos are now deactivated.

It also unifies the use of Nanos and Nanox. The nade of the Library is Nanos++,
whereas the previos version was Nanos4. In order to have a unique number,
we use "nanos" for all reference in rose, and "nanox" only for call
to the actual Nanos++ library.