(CMAKE ROSE) fix crashing tests
In a previous commit, I reorganized the binary directory so that all
executables are put in the "bin" directory, and all libraries are put
in the "lib" directory.  This change caused two of the tests to start
crashing.  This is due to the fact that ROSE now thinks it's being run
from an install tree, even though it is actually still running from a
build tree.  When running from an install tree, ROSE expects to find
its compiler-specific headers in $ROSE-install/include.  When it's
running from a build tree, ROSE instead looks in $ROSE-build/include-staging.
This difference led to the previously mentioned segmentation fault.

To work around this problem, CMake now creates a symbolic link from
$ROSE_BIN/include to $ROSE_BIN/include-staging