Fix OpenMP outlining of arrays from function scope
Arrays from function scope that are private must be passed as copy,
since the pointer must not be modified, but the pointed data must be shared.
This means that these variables must not be shared in the outlined struct
(no additional pointer for them) but must be used as shared when unpacking
(catch the address of the variable) and in the outlined function (replace
all uses by a pointer reference to the variable).