Add support for multidimensional shapping expressions
This commit extends the support for shapping expressions with more than one dimension.
Add new tests for this feature as well.

The commit also fixes an error when patching firstprivate variables: shaping expressions and
array sections appearing in a task depend clause are considered are outlined as firstprivate
instead of shared (as happens to any other variable appearing in a depend clause)