Include Shape Expressions and Array Sections as Expressions
This expressions are allowed in OpenMP task depend clauses.
This commit replaces the previosus support for this kind of expressions in the depend clause.
Map clause still works with the previous implementation of array sections,
which creates a map <symbol, vector<sections>. This implementation is not correct for directives
using the same variables with and without sections, since the array only links a symbol with
a section, but dosen't link an specific instance of the variable. Example: